+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v3.15.0
 ## 🚀 Feature
 - `AndesCarousel` supports XSmall margin and custom horizontal padding
+- Added Checkbox support to `AndesTextfield` | Author: [@MKiperszmid](https://github.com/MKiperszmid)
 ## 🛠 Fixes
 - Fix `AndesCheckbox` overlap of layouts, set ellipsize end and the support of multi lines | Author: [@mingkao-meli](https://github.com/mingkao-meli)